refactor: reorganized file structure for questionnaire builder components
