+++ app/(dashboard)/superadmin/questionnaires/preview/page.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 
-import { QuestionnairePreviewRenderer } from "@/components/faculytics/questionnaires/questionnaire-preview-renderer";
+import { QuestionnairePreviewRenderer } from "@/components/faculytics/questionnaires/builder/questionnaire-preview-renderer";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { deserializeQuestionnaireVersionToDraft } from "@/lib/questionnaires/builder-deserializer";
