+++ components/faculytics/questionnaires/builder/questionnaire-qualitative-editor.tsx
@@ -2,7 +2,7 @@
 
 import { Trash2 } from "lucide-react";
 
-import { QuestionnaireAddActionButton } from "@/components/faculytics/questionnaires/questionnaire-add-action-button";
+import { QuestionnaireAddActionButton } from "@/components/faculytics/questionnaires/builder/questionnaire-add-action-button";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Checkbox } from "@/components/ui/checkbox";
